google_adk_Comprehesenive_usecase/pyproject.toml
[project]
name = "google_adk_workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.3.0",
    "litellm>=1.67.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.7",
]